requirements.txt
colour
numpy
Pillow
scipy
tqdm
pycairo
pydub
pygments
pyreadline; sys_platform == 'win32'
rich>=4.2.1
# pytest (for development only)
attrs~=19.3.0
# black (for development only)
